Group Policy DB-2: Classifiers, Actions, and Rules
(Patch series identifier: GP-DB-2)

This patch includes the DB layer for Policy Classifiers, Policy Actions, and
Policy Rules resources.

In the context of the larger Group Policy Model, the Policy Rule resource
is referenced by Contracts, which be will introduced in a subsequent patch.

The Plugin layer to back these resources will also be introduced in a subsequent patch.

Partially-implements: blueprint group-based-policy-abstraction

Change-Id: I6ae9047539c401b2403ca0d4fe010cad9a97cfdc
Author: Sumit Naiksatam <sumitnaiksatam@gmail.com>
Co-Authored-By: Bob Kukura <kukura@noironetworks.com>
Co-Authored-By: Stephen Wong <s3wong@midokura.com>
Co-Authored-By: Mohammad Banikazemi <mb@us.ibm.com>
Co-Authored-By: Mandeep Dhami <dhami@noironetworks.com>
Co-Authored-By: Ivar Lazzaro <ivarlazzaro@gmail.com>