Group Policy API-1: EP, EPG, L2 Policy, L3 Policy
(Patch series identifier: GP-API-1)
This is the first in a series of patches which implement Group Policy. It
includes the Resource Model and API for Endpoint, Endpoint Group, L2
Policy, and L3 Policy resources.
In the context of the larger Group Policy Model, the Endpoint Group resource
references Contracts, which will be introduced in a subsequent patch.
The DB and Plugin layers to back these resources will also be introduced in
subsequent patches.

DocImpact
Gerrit Spec: https://review.openstack.org/#/c/89469

Change-Id: I4f4cc7e2a899b39947b784ec390c3df599cf01ae