Refactor GBP resource mapping with Neutron RESTful APIs
1. Decouple GBP resource mapping from neutron plugins and use
Neutron RESTful APIs instead.

2. Move neutron_client related calls to neutron_api_mixin for
easier removing in the future.

3. Add mock_neutronv2_api to provide methods to patch the
Neutron RESTful API calls with WSGI calls.

4. Modify the existing 'UT' for the resource mapping driver, by
patching the RESTful APIs with the WSGI calls, to ensure the
test scenarios are still examined and passed.

5. Modify the existing 'UT' for the APIC mapping driver, by
patching the RESTful APIs with the WSGI calls, to ensure the
test scenarios are still examined and passed.

Change-Id: I9ec6d159b87e77fcd034b2ff6d4e5c0969798b34
Author: Yi Yang <yyos1999@gmail.com>
Co-Authored-By: Yapeng Wu <yapengwu@gmail.com>
Co-Authored-By: Ivar Lazzaro <ivarlazzaro@gmail.com>
Co-Authored-By: Robert Kukura <kukura@noironetworks.com>