Refactor GBP resource mapping with Neutron RESTful APIs
1. Decouple GBP resource mapping from neutron plugins and use
Neutron RESTful APIs instead.

2. Add mock_neutronv2_api to provide methods to patch the
Neutron RESTful API calls with WSGI calls.

3. Modify the existing 'UT' for the resource mapping driver, by
patching the RESTful APIs with the WSGI calls, to ensure the
test scenarios are still examined and passed.

4. Modify the existing 'UT' for the APIC mapping driver, by
patching the RESTful APIs with the WSGI calls, to ensure the
test scenarios are still examined and passed.

5. Modify the existing 'UT' for the One Convergebce GBP driver,
by patching the RESTful APIs with the WSGI calls, to ensure the
test scenarios are still examined and passed.

Change-Id: I0626359d637246f0252c46c9a8b6c80317d24bc2
Signed-off-by: Yi Yang <yyos1999@gmail.com>
Author: Yi Yang <yyos1999@gmail.com>
Co-Authored-By: Yapeng Wu <yapengwu@gmail.com>
Co-Authored-By: Ivar Lazzaro <ivarlazzaro@gmail.com>
Co-Authored-By: Robert Kukura <kukura@noironetworks.com>