Refactor GBP resource mapping with Neutron RESTful APIs
1. Decouple GBP resource mapping from neutron plugins and use
Neutron RESTful APIs intead.

2. Add moc_neutronv2_api to provide methods to patch the
neutron RESTful API calls with direct neutron-plugin API calls.

3. Modify the existing 'UT' for the resource mapping driver, by
patching the RESTful APIs with the neutron-plugin API calls, to
ensure the test scenarioes are still examed and passed.

4. Modify the existing 'UT' for the APIC mapping driver, by
patching the RESTful APIs with the neutron-plugin API calls, to
ensure the test scenarioes are still examed and passed.

5. Modify the existing 'UT' for the OneConvergebce GBP driver,
by patching the RESTful APIs with the neutron-plugin API calls,
to ensure the test scenarioes are still examed and passed.

Change-Id: I0626359d637246f0252c46c9a8b6c80317d24bc2
Signed-off-by: Yi Yang <yyos1999@gmail.com>
Author: Yi Yang <yyos1999@gmail.com>
Co-Authored-By: Yapeng Wu <yapengwu@gmail.com>