Adding GBP integration and functional gate
This patch implements the hooks so that functional and integration
tests can be run in the group-based-policy repo.

Currently, the configured devstack for this job is enabled to run GBP
exercise scripts, which result in complete end-to-end integration
testing of the GBP service and client components along with integration
with Neutron and Nova.

A placeholder functional test case is also added. A functional test suite
will be added in a separate commit.

Change-Id: Id705e5755b9662f6e297f29836d6983d66f845a8