Properly handle driver exceptions during postcommit
If any exception is raised by a driver's create_<resource>_postcommit
method, the partially-created resource is deleted before returning the
exception to the client.

If any exception is raised by a driver's delete_<resource>_postcommit
method, the exception is logged, but is not returned to the client.

Also, the log messages are made more consistent and readable.

Closes-bug: 1407321
Partial-bug: 1416177
Change-Id: I7c3096ae436f23d68d695b813f7d8d0b2588dbfe