[WIP] reorganize service chain directory
With the introduction of the new servicechain plugin, some sort
of path reorganization is required so that the code doesn't become
confusing as new plugins are implemented.

The existing servicechain_plugin has been renamed "MSC" for
Modular Service Chain. The name recalls ML2 because of the MSC
structure being inspired by it.

The old entry point name is not removed from setup.cfg for backward
compatibility, still need to figure out a way to rename the configuration
files properly having the same goal in mind.

Partially implements blueprint node-centric-chain-plugin

Change-Id: I4f8db6f5fa30479aad283b07e499901af000a6c5