Fix PTG update with a different NSP of same type(self_subnet)
On PTG update with a new NSP, first we do an update on PTG subnet
to restore the IP reserved for NSP back to the PTG subnet as part
of clean up of first NSP. Again we do an update subnet allocation
pool as part of setting the new NSP. This update subnet sometimes
does not see the updated subnet object. For now expunging all
objects in the current session t fix this.

Change-Id: Ied854a5fbd7eaf95ffcf85e9be5ada2bc905d723
Closes-bug: 1426946