subnet allocation improvement
This patch removes some code duplication on L3P pool subnet allocation
and improves performances removing linear search in most
cases. This is not a final solution for subnet allocation, which
is deferred until we support Neutron's subnet pools, but it's a
requirement for the upcoming TScP work.

Partial-bug: 1382149
Partially implements blueprint node-centric-chain-plugin

Change-Id: I1ef8b377b3eee03158a0776406af4aa6ab27f5c8