[WIP] proxy ptg extension
Driver extension that allows users to specify a 'proxied_group_id'
during PTG creation. Whenever a PTG proxies another group, it is
expected to intercept all the traffic before sending it to the
original destination. Proxy groups can modify the traffic if needed.
This is a useful construct for the Traffic Stitching Plumber.

Partially implements blueprint node-centric-chain-plugin

Change-Id: Idc185df5b0c7e61ef800ca449911656a8c1d2b87