explicitly specify router_id when creating a floating ip
As for today, Neutron requires a router with an external
interface to be directly attached to any internal subnet
in which the user wishes to instantiate a floating IP.

With this small patch, we experiment a work around that allows
internal calls (ie. directly issued by the L3 plugin) to explicitly
specify a router ID without any checking occurring on
its correctness. When using this, the developer needs
to make sure that the specified router can actually handle
NAT properly.

Partially implements blueprint node-centric-chain-plugin

Change-Id: I6bd63c34cacd65a7fbfca66b8e05649e96187b1e