Adding metadata server and DNS SG default rules
Since we started filtering at the egress, we explicitly
created allow rules only for the PRS consumed by a
group. However, we also need to open up metadata server
and DNS access.

In this patch we update the default security group with
a rule to allow TCP access to 169.254.0.0/16 and, TCP and
UDP access to DNS port 53 (for both ipv4 & ipv6).

Change-Id: I5a0e329a7491abbcae0de1a2908325e0159138a8
Closes-bug: 1487156