Add support for NAT to APIC mapping driver
To support NAT-ing, several changes are made to external-segment
handling in the APIC mapping driver. The basic idea is that
translated IPs (assigned floating-IPs or host-specific IPs for
SNAT) appear to be in a new endpoint-group called NAT EPG.
In addition, shadow objects are created for L3-Outs and external
EPGs whenever an L3-policy is associated with an external-segment.
These shadow objects provide/consume the contracts that the
user has specified with the external-policy/external-segment.
The real/non-shadow objects in ACI provide/consume an allow-everything
contract with the NAT EPG also providing/consuming the same
contract. Also, any APIC configuration for the L3Out (e.g. logical
node profile) remains only with the non-shadow L3Out.

With the use of shadow L3Outs, we are able to remove restriction
that an external-segment can have only one L3-policy.

This change also enables auto-allocating floating IPs to policy
targets by creating the necessary plumbing in Neutron (creating
external network, creating router and plugging it to right
subnets).

Closes-bug: 1494023
Change-Id: I13994bcddf8f5c93bbe348d5c22d933d3fa9f143
Signed-off-by: Amit Bose <amitbose@gmail.com>