WIP: Handle concurrent implicit L3P creation
Raise DefaultL3PolicyAlreadyExists if a policy named 'default' already
exists for the tenant when creatig an L3 policy.

If other threads or processes concurrently trying to create a default
L3 policy for a tenant results in a DefaultL3PolicyAlreadyExists
exception, use the default L3 policy created by the thread that
succeeded.

Fix RMD's delete_policy_target_group_postcommit to gracefully handle
partially constructed states, such as a null L2 policy or a subnet
not attached to a router.

Currently experimenting with rally tests.

Closes-bug: 1462024

Change-Id: I09f29eef22edb45290070aae30e97c93c77ea341