Handle NAT-pools properly in APIC-mapping driver
For every NAT-pool associated with an external-segment,
create a subnet in APIC in the corresponding NAT BD. This
is needed so that ACI recognizes traffic using floating-IP
as being valid. Also, enforce that any created  NAT pools
do not overlap with APIC external network ('cidr_exposed'
in config file) or the host-pool for SNAT. As a related
change, have tied the lifecycle of NAT EPG/BD/VRF to that
of the external segment in GBP to prevent stale objects
being left behind in APIC.
Finally, when floating-IPs are allocated manually by
user, make sure that the allocation comes from the
associated NAT pool. Factored out the FIP allocation
code in resource-mapping driver for this.

Bonus fix: stringify 'vrf_name' when responding to RPC

Change-Id: I43ce97ae651ac5683809edf26f9dbe1beac18276
Signed-off-by: Amit Bose <amitbose@gmail.com>