Monkey patch to allow setting router's external IP
In Juno, supplying a specific subnet/IP when setting a
router's gateway is ignored. As a result of this,
routers created for an external segment many not be
assigned addresses in the subnet associated with that
external segment, specially if there are nat-pools for
that external segment.
This behavior was fixed in Kilo through commit
66f5eadec0adfa1dc55809fbe52b087a17c8a794 ('Allow setting
a tenant router's external IP'). This monkey patch is a
slightly modified version of the code in the above commit.

Closes-Bug: 1499053

Change-Id: I68f1cb2f620c9211dddc53dd4611a283160a3f6d
Signed-off-by: Amit Bose <amitbose@gmail.com>