Update ODL Mapping
Previously, Endpoint Group only works with a single named selector.
Recently, with updated yangtools, EPG requires a list of named
selector. This patch is to support such a change.

Also, because EGP only supported a single selector, a tweak had to
be done to map multiple policy rule sets to a single contract. Now,
as EPG supports a list of selectors, each policy rule set is mapped
to one contract.

1. Update odl_mapping
2. Update UT for odl_mapping

Change-Id: Ie3fe0ded1cfa846a703c02e9d3864be49c3a8ed0
Signed-off-by: Yi Yang <yyos1999@gmail.com>