[apic-mapping] Host SNAT IPs from user subnets
The host SNAT auto IP allocation feature creates a SNAT subnet
on a SNAT network based on the CIDR in the host_pool_cidr configuration.
However, if this CIDR is exhausted, there is no way to add more subnet(s).

With this patch, the user will be able to create additional subnet(s) on
the external network associated with the external segment. The subnet has
to be named "host-snat-pool-for-internal-use". Once this subnet is created
successfully, the host SNAT IPs will continue to be allocated from the newly
added subnet(s).

Change-Id: I1549e9ad5cef74e91eb108811514233da0bf0dfc
Closes-bug: 1518520