Avoid allocating multiple router gateway IPs
When an L3Policy is implicitly associated with the
'default' external-segment, multiple IP addresses
get allocated for the gateway port. This happens
because update-postcommit is called before
create-postcommit during the creation of L3Policy.
This change avoids the update-postcommit;
instead we set things up in the DB and request
context in a way so that create-postcommit behaves
as if the 'default' external-segment was specified
as part of the create request.

Closes-Bug: 1522554

Change-Id: Iebea8cf315d2ed7acbf4994940ccd57a93133475
Signed-off-by: Amit Bose <bose@noironetworks.com>
(cherry picked from commit 46a121339ffee90154cfc900a8991735746a4c11)
(cherry picked from commit 2b370862fa835977bfb109899928503c2f6c4fad)