Allow disabling of default route via DHCP
By default the route to the default gw is always injected via DHCP.
This patch adds an attribute "inject_default_route" to the l2_policy
which allows controlling this behavior. The default behavior is still
maintained.

The apic_mapping driver is updated to support both setting and unsetting
this attribute.

The resource_mapping driver currently only supports setting this attribute.

Implements blueprint inject-default-route

Change-Id: Ifdf6d51ae3a115f3d643b74e12d5693be3d15307