[APIC mapping] Send correct host SNAT IP in get_gbp_details
We look at the binding-host of a port to decide the
host SNAT IP that should be returned in GBP details.
For a VM that is in the process of live migration,
the binding-host may still be the source host while
the RPC request comes from the destination host. In
this case we wrongly send the host SNAT IP of the
source host to the destination host.
The fix is to use the host of the agent making the
RPC call to decide which host SNAT IP to report.

Closes-Bug: 1523633

Change-Id: I53c8fdda8c377a94a745955c753f9c0e8233bb5e
Signed-off-by: Amit Bose <bose@noironetworks.com>
(cherry picked from commit f2c3b1478d6cb4fc3e9be8be35894ca557e401bc)