port extra attribute extension for gbp mapping
Just a few lines of code to propose a way for internal users to
specify extra port attributes when creating a PT.
There's no persistence or API changes involved, and the workflow
will look like the "implicit port" workflow. The advantage of
having this is that some tweaks can be made to the Neutron port
by still having GBP as the port owner.

Closes-Bug: 1527360
Change-Id: Iebcda544d0b5aee3e02e4828228d662997008251