port extra attribute extension for gbp mapping
Just a few lines of code to propose a way for admins to specify
extra port attributes when creating a PT. There's no persistence
involved, and the workflow will look like the "implicit port"
workflow. The advantage of having this is that some tweaks can
be made to the Neutron port by still having GBP as the port
owner.

Implements blueprint port-extra-attributes
Change-Id: Iebcda544d0b5aee3e02e4828228d662997008251