[APIC mapping] Fix contract association for pre-existing external EPG
When any tenant that uses a pre-existing L3Out (and external EPG),
either deletes its external policy (or detaches it from
external-segment), we were removing the provide/consume
relationship for the external EPG to NAT-allow contract.
This breaks NAT connectivity for the other tenants. This
fix removes the relationship only if there are no other
external policy objects mapping to the same external EPG.

Closes-Bug: 1527763

Change-Id: I74858c83b111b1fb223e4cd6af87e3625c64ec49
Signed-off-by: Amit Bose <bose@noironetworks.com>