Initializing proxy pool attrs from conf
We had defined the following configuration:
[group_policy_implicit_policy]
...
default_proxy_ip_pool=<cidr>
default_proxy_subnet_prefix_length=<integer>
...

but this was being used only when a L3P was being created
implicitly. Per the bug filed it is desirable to pick up
these config values even when a L3P is explicitly created
(and the proxy group is created internally).

This is now achieved by moving the above config to a new
section:
[group_policy_proxy_group]

This configuration is loaded from the proxy_group externsion
is and the default values are set from the loaded configuration.

And while at it, the default_proxy_subnet_prefix_length is being
increased to 28 to accomodate basic cases for services with
mutiple interfaces, and/or in HA configuration.

Change-Id: Ib955611cbf4516e7502d2b07e5b44589dc95b325
Closes-bug: bug/1535649
(cherry picked from commit d164bdcc46f83e0643451de8d2f9d7c7f60a8825)