Adding dir structure for developer reference docs
This patch is in response to prior discussion in the
weekly IRC team meeting.

The patch does not actually add any documentation (apart
from the bits on testing and dev env setup). It does propose
topics in the index.rst for documents that need to added for
the existing components. These will be added by the team as
follow-up patches.

For any patch that adds a new feature or modifies an existing
feature it will be required to submit a devref rst document
as a part of that patch. This document should be posted in the
doc/source/devref/ location. There isn's a specific format for
the document, but it is expected that the document adequately
explains the design of the feature that is being implemented
such that its easy for the rest of the team to review, maintain,
and enhance the implementation.

Change-Id: Ifc32c330d754f623c403fb0c9a5f5c639cce41db