Notify DHCP agent on Router interface port delete
The traffic stitching implementation deletes the router interface
and creates a Service Target with the gateway IP for the Service
VM to be brought up as the gateway for the provider. The DHCP
agent actually has the router port details in the DHCP hosts file.
The DHCP agent is not receiving a port delete notification from
either core plugin or L3 plugin for router interface port delete.
Hence it ends up having two entries for same IP in the hosts file.
Because of this the Service VM does not get an IP through DHCP.

As a fix, delete port notification is sent after invoking the
remove router interface api.

Change-Id: I485a2eea573bb318cd70c13a2600ffa318778a2b
Closes-Bug: 1552176
(cherry picked from commit 588a65ad904ab037ed8f1f5a6fc3c7e006314538)