Fix L3 APIC plugin inheritance
This removes the NeutronDbPluginV2 class from the
inheritance hierarchy, and changes from the DVR
parent class to the non-DVR class. It replaces
references to the ML2 core plugin with the member
variable defined in the superclass.  It also adds
unit tests for the plugin.

Change-Id: I99d3341e17aef320767515c52db77902e72a68c6
Closes-Bug: 1555412
(cherry picked from commit 23148a51e26c790510898be78d1270d069c9aabe)