[APIC-mapping] Fixes for pre-existing L3-out
Fix #1: When verifying that the tenant of pre-existing
L3-out is correct, comparison of ApicName objects was
failing incorrectly. Fix is to compare the string
representation of ApicName-s.
Fix #2: When L3policy is associated with pre-existing
NAT-disabled L3-out, earlier we would reject the call if
the VRF mapping to L3policy wasn't already associated
to the pre-existing L3-out. This forces an awkward
workflow whereby the pre-existing L3-out needs to be
updated manually in ACI after the VRF has been created.
In this change, we update the pre-existing L3-out
automatically instead of throwing an error.

Bonus fix: Latest ApicManager implicitly parses config
files to build switch dict, external network dict etc.
So these can be avoided in the APIC mapping driver.

Change-Id: I688e31bb2f42964a3799e0f0febc9f46864cd5c9
Signed-off-by: Amit Bose <amitbose@gmail.com>