Refactor for create_floatingip call
The current implementation for create_floatingip results in the
L3 API being called twice. The first pass causes additional attributes
to be populated, which are used in the second call. This patch converts
this to a single call, which gets the parameters needed and continues,
rather than call itself recursively.

Change-Id: Ie9a3a57d2186fd693e9bc1e03501f0038e316b41
Closes-Bug: #1561638