[WIP] New APIC mechanism and extension drivers
This is a very preliminary work-in-progress version of a new APIC
mechanism driver utilizing the ACI Integration Module (AIM) library
concurrently being developed. A correspinding extension driver exposes
details regarding the mapping of the Neutron resources to APIC. These
drivers require the Ml2Plus extended driver APIs.

Currently, only the tenant and network name mapping and the AIM
integration for bridge domains are implemented. The
synchronization_status attribute is not yet implemented in the
extension driver. Units tests are also insufficient.

Change-Id: I82df32f0880d6a0d53b305f6c6391fcbea049d1b