Quality of Service support (PoC 01 - PTG NSP)
QoS support for PTs inside PTG, as a new NSP param type.

The new NSP param type, qos_maxrate, maps to the Neutron QoS Policy
resource, which then gets associated to the QoS bandwith limit rule
with a certain maximum rate (in Kbps) set. Support for the burst rate
is out of scope in this initial code for PoC 01.

Change-Id: I3950a13c482d7a7e1fa03667a543628aaa36ee6e
Implements: blueprint initial-support-qos