Quality of Service support via NSPs
QoS support for PTs inside PTG, as 2 new NSP param types.

The new NSP param types, qos_maxrate and qos_burstrate, map to the
Neutron QoS Policy resource, which then gets associated to a QoS
bandwith limit rule with a certain maximum rate and/or burst rate (in
Kbps and Kbs respectively) set.

This is still a WIP, the following will be done in further patchsets:
- Add exercise script to enable integration testing and workflow doc;
- Fix integration gate job failure, if applicable;
- Add DevRef regarding this patch.

Change-Id: I3950a13c482d7a7e1fa03667a543628aaa36ee6e
Implements: blueprint initial-support-qos