[APIC-mapping] Support non-OpFlex networks
This change enables the apic-mapping driver to
handle endpoints (ports) when the network-type
is VLAN. In this scenario, there are no OpFlex
agents on the compute/network nodes that can configure
the datapath to isolate traffic and enforce policy.
Instead we use additional 'shadow' Neutron networks,
one per PTG, to isolate ports within a PTG from other
PTGs. Static path binding is configured on-demand on
APIC using the segmentation-ID of the shadow networks.

These additional "shadow" networks mirror the
configuration of the Neutron network created for
the L2-policy of the PTGs:
- Subnets in the L2P network are reflected in shadow
networks of each PTG associated with the L2P
- Implicit ports created for policy-targets are mirrored
in the shadow network of the corresponding PTG. This
additional 'shadow' port is now associated with the
PT instead of the implicitly created port.
- Explicit ports for PTs is allowed as long as they have
been created in the appropriate shadow network
- DHCP ports on the shadow networks are disabled to avoid
unwanted/wrong responses to DHCP requests from VMs.

This scheme imposes a few restrictions:
- PTG of a PT can be no longer updated because that would
require changing the network of the associated port
- Ports created in the L2P network cannot be associated
with a PT; instead such an explicit port must be created
in the shadow network for the PTG of the PT.

Without OpFlex agent, NAT-ing won't work with APIC. So
the apic-mapping driver configures external-segments on
APIC for the no-NAT case. Once support for edge-NAT is
available (e.g. through ASR integration), we can make
the appropriate changes to make floating-IPs and PAT
work.

Change-Id: I39d09257f6eadcf5ae527d779f13493415a4ec85
Signed-off-by: Amit Bose <amitbose@gmail.com>