Remove syslog conf from devstack gate job
Syslog logging is not required in the gate job.

Also the LOG_COLOR is set to True by default but
is causing the console log to be inundated with:
tput: No value for $TERM and no -T specified

so, setting it to False.

Change-Id: I6faf9c73bd3fb273861ed2d4bfc9bf0f6d6be147
(cherry picked from commit 5bbcf15e45cf509240555c1066261580a5a25ee2)