Fix Port Group name construction
The port group name was improperly constructed when used
with single_tenant_mode = False. It also wasn't passing
the port group name into the RPC call to the DVS agent.
Lastly, the port needs to be marked as active as part of
the port binding process (DVS ports only). This patch
fixes all of these issues.

Closes-Bug: #1590609, #1590830

Change-Id: I104fc8d24179a673f4d7f3df5ec6e5e3369c5d9a
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>