Better quota enforcement for GBP resources
This patch aligns the quota enforcement in GBP with the
changes made in Neutron quota management as a part of the
following blueprint and implementation patches referenced
therein:
https://blueprints.launchpad.net/neutron/+spec/better-quotas

The decorator @resource_registry.tracked_resources  in the
neutron.quota.resource_registry module is applied to the GBP
and Service Chain plugin __init__ methods.

In addition, the local_api module is also enhanced to enforce
quotas when GBP and Service Chain resources are created implicitly.

Older UTs that test the quotas have been preserved thus validating
the approach (they have only been moved to the appropriate test classes
on account of the resource tracking logic having moved from the earlier
DB modules to the plugin modules).

Closes-bug: 1594215

Change-Id: I38093426c89d1f4398c40da1f3392c0988831ff8