Splitting NFP integration tests into new job
Adding the necessary hooks to setup and run the NFP integration
tests.

A complementary patch will be submitted in the OpenStack project-config
to define a new job, but this patch will have to be merged for that
patch to merge.

Once the above two patches are merged, a follow up patch will be submitted
to remove running the NFP integration tests from the current integration
job.

Change-Id: I73c1c76b2bb8bc20fa42b342b01e2ce61da420d3
(cherry picked from commit a002cc58ec9fb4ffecc6f773798e19015000b6d1)