AIM Policy Driver - Part 2 - Extension Driver
This adds a AIM specific extension to reflect the APIC DN for
a GBP resource that maps to an AIM resource. This patch implements
this only for the PTG resource.

This also sets the status of the PTG based on the AIM EPG status.

This also updates the devstack setup to include the aim_mapping
GBP policy driver configuration.

Change-Id: I30f5e5e63b3b172eb79c8a9934eb662928d13f6c