[WIP] AIM Policy Driver - Part 3 - Policy Resources
Policy Rule is mapped to an AIM Filter, and the Policy Classifier
is mapped to an AIM FilterEntry.

This also fixes a bug in the _get_status_from_drivers() logic which resulted
in the driver extension attributes being dropped from the result that was
returned. This allowed validating the "apic:distinguished_names".

Change-Id: I7da7ecedd33a4c38623c944e366fbc01d216fdfa