AIM Policy Driver - Part 3 - Policy Rules
Policy Rule is mapped to an AIM Filter, and the Policy Classifier
is mapped to an AIM FilterEntry.

The older apic_mapping driver has been refactored slightly to create
a new library of functions called apic_mapping_lib that can be used both
from the apic_mapping driver as well as the new aim_mapping driver.

The rollback testing for the aim_driver has been enhanced by adding the
dummy_driver as second policy driver for testing, and having the dummy_driver
raise exception to trigger the rollback. This ensures that the entire code in
the aim_driver is executed before the exception, and hence the complete
rollback is tested.

This also fixes a bug in the _get_status_from_drivers() logic which resulted
in the driver extension attributes being dropped from the result that was
returned. This fix now allows validating the "apic:distinguished_names".

Change-Id: I7da7ecedd33a4c38623c944e366fbc01d216fdfa