[apic_aim] [WIP] Map neutron resources to AIM, part 2
Neutron address scopes are mapped to AIM VRFs, with the DN and sync
status available via extended attributes. Neutron networks are
initially created with BDs that have routing disabled and are
associated with a common unrouted VRF.

Tracking the relationship between address scopes, subnetpools,
subnets, networks and routers is not yet implemented.

An obvious bug in sync_state handling is also fixed, but a unit test
to prevent regressions is still needed. The devstack plugin is also
fixed to include the proxy_group extension driver.

This version is definitely a work-in-progress, not meant to be
merged. It is intended to support testing of the control and data
paths for individual unrouted networks only. Mapping of subnets is
disabled, pending moving that to when they are added as router
interfaces. Address scopes are mapped to VRFs, but these VRFs are not
yet associatied with networks' BDs.

Change-Id: I8fef7809654924d4c3ff50bdf49188faba0684e0