[apic_aim] [WIP] Map address scope to AIM VRF
Neutron address scopes are mapped to AIM VRFs, with the DN and sync
status available via extended attributes.

Tracking the relationship between address scopes and other Neutron
resources, such as networks, subnets, and routers, and returning the
mapped VRFs via RPCs, are not yet implemented.

An obvious bug in sync_state handling is also fixed, but a unit test
to prevent regressions is still needed.

Change-Id: I8fef7809654924d4c3ff50bdf49188faba0684e0