AIM Policy Driver - Part 4 - Policy Rule Sets
Policy Rule Set is mapped to an AIM Contract and ContractSubject. All
AIM Filters mapped to the Policy Rules that are part of the PRS, are
added to the in, out, or bi filter lists of the ContractSubject depending
on the direction of the Policy Classifier of the Policy Rule.

The PTG UTs have been updated to include testing for setting of provided
and consumed PRSs.

Some additional refactoring has been done to address comments from the
previously merged patches.

Change-Id: Ic147538b43564483e60fd6308bb821f4846466f8