AIM Policy Driver - Part 4 - Policy Rule Sets
Policy Rule Set is mapped to an AIM Contract and ContractSubject. All
AIM Filters mapped to the Policy Rules that are part of the PRS, are
added to the in, out, or bi filter lists of the ContractSubject depending
on the direction of the Policy Classifier of the Policy Rule.

The L2P creation has been updated to add Infra Service and ARP contracts
to be provided/consumed by the default EPG for the L2P.

The Infra Service and ARP contracts are created once per tenant when the first
L2P is created. Every subsequent L2P's default EPG in that tenant uses the same
Infra Service and ARP contracts. When the last L2P is deleted, these Infra
Service and ARP contracts are deleted too.

Every new AIM EPG created consumes the Infra Service and ARP contracts.

The PTG UTs have been updated to include testing for setting of provided
and consumed PRSs.

display_name is being set for all AIM resources (currently, same as
corresponding GBP resource name).

The extension attribute handling for the PTG in the AIM driver is updated
to not fetch from the DB.

DB table clean up is added for AIM test module.

Some additional refactoring has been done to address comments from the
previously merged patches.

Change-Id: Ic147538b43564483e60fd6308bb821f4846466f8
(cherry picked from commit 10110aae8b3ab1e9596e9c6c2129e3831f4888e8)